update TextArea stories to latest recommendations
